requirements.txt
pandas
numpy
scikit-learn
tensorflow
requests
matplotlib
python-dotenv
streamlit 
plotly
